Image-caption-generator/requirements.txt
streamlit==1.33.0
transformers==4.40.0
torch==2.2.0+cpu --index-url https://download.pytorch.org/whl/cpu
Pillow==10.3.0
gTTS==2.5.1
protobuf==3.20.3

